% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.gg}
\alias{dir.gg}
\title{Concatenates a path, filename, and file extension into an output path, then writes a plot to the output path.
Any folders in the path that do not exist are created.}
\usage{
dir.gg(x, filename, path = ".", ext = "pdf", ..., append.date = T)
}
\arguments{
\item{x}{A plot object.}

\item{filename}{The output file name.}

\item{path}{The path to the output file.}

\item{ext}{The file extension.}

\item{...}{Additional arguments to \code{ggplot2::ggsave}.}

\item{append.date}{Whether to write the output to a directory with today's date.}
}
\description{
Concatenates a path, filename, and file extension into an output path, then writes a plot to the output path.
Any folders in the path that do not exist are created.
}
\examples{
library(ggplot2)
x <- ggplot2(iris)
dir.gg(x,"example",path="path/to")
plot(1:5,1:5)
dev.off()
}
\seealso{
\code{\link{dir.img}}, \code{\link{ggplot2::ggsave}}
}
