% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.hist}
\alias{dir.hist}
\title{Accepts a numeric vector and writes a histogram in EPS format to a specified file.
Allows for setting all values greater than a given quantile to the quantile value.
Any folders in the path that do not exist are created.}
\usage{
dir.hist(
  x,
  file,
  breaks = "FD",
  quant = 1,
  path = ".",
  append.date = T,
  col = "gray28",
  border = F,
  main = "",
  xlab = "length (bp)",
  ...
)
}
\arguments{
\item{x}{A numeric vector.}

\item{breaks}{A \code{breaks} argument passed to \code{hist}.}

\item{quant}{The maximum quantile to be plotted.}

\item{path}{The path to the output file.}

\item{append.date}{Whether to write the output to a directory with today's date.}

\item{...}{Additional arguments to  \code{hist}.}

\item{filename}{The output file name.}
}
\description{
Accepts a numeric vector and writes a histogram in EPS format to a specified file.
Allows for setting all values greater than a given quantile to the quantile value.
Any folders in the path that do not exist are created.
}
\examples{
dir.hist(1:100,"example",path="path/to",quant=.9)
plot(1:5,1:5)
dev.off()
}
\seealso{
\code{\link{dir.eps}}, \code{\link{hist}}
}
