% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.img}
\alias{dir.img}
\title{Concatenates a path, filename, and file extension into an output path, then opens a connection to the output path using function \code{fn}.
Any folders in the path that do not exist are created.}
\usage{
dir.img(filename, fn, ext, path = ".", ..., append.date = T)
}
\arguments{
\item{filename}{The output file name.}

\item{fn}{Function that accepts x as its first argument and a file path as its second argument.}

\item{ext}{The file extension.}

\item{path}{The path to the output file.}

\item{...}{Additional arguments to \code{fn}.}

\item{append.date}{Whether to write the output to a directory with today's date.}
}
\value{
The resulting value of \code{fn}.
}
\description{
Concatenates a path, filename, and file extension into an output path, then opens a connection to the output path using function \code{fn}.
Any folders in the path that do not exist are created.
}
\examples{
dir.img(pdf,"example","pdf",path="path/to")
plot(1:5,1:5)
dev.off()
}
