% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.out}
\alias{dir.out}
\title{Concatenates a path, filename, and file extension into an output path, then writes data \code{x} to the output path using function \code{fn}.
Any folders in the path that do not exist are created.}
\usage{
dir.out(x, fn, filename, ext = "txt", path = ".", ..., append.date = T)
}
\arguments{
\item{x}{Data to be written to filename.}

\item{fn}{Function that accepts x as its first argument and a file path as its second argument.}

\item{filename}{The output file name.}

\item{ext}{The file extension.}

\item{path}{The path to the output file.}

\item{...}{Additional arguments to \code{fn}.}

\item{append.date}{Whether to write the output to a directory with today's date.}
}
\value{
The resulting value of \code{fn}.
}
\description{
Concatenates a path, filename, and file extension into an output path, then writes data \code{x} to the output path using function \code{fn}.
Any folders in the path that do not exist are created.
}
\examples{
dir.out(iris, write.table,"example","txt",path="path/to")
}
