% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{dir.png}
\alias{dir.png}
\title{Concatenates a path, filename, and file extension into an output path, then opens a png connection at the output path.
Any folders in the path that do not exist are created.}
\usage{
dir.png(filename, path = ".", ..., append.date = T)
}
\arguments{
\item{filename}{The output file name.}

\item{path}{The path to the output file.}

\item{...}{Additional arguments to \code{png}.}

\item{append.date}{Whether to write the output to a directory with today's date.}

\item{res}{The nominal resolution in ppi.}

\item{width}{The width of the device.}

\item{height}{The height of the device.}
}
\description{
Concatenates a path, filename, and file extension into an output path, then opens a png connection at the output path.
Any folders in the path that do not exist are created.
}
\examples{
dir.png("example",path="path/to")
plot(1:5,1:5)
dev.off()
}
\seealso{
\code{\link{dir.img}}, \code{\link{grDevices::png}}
}
