% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirfns.R
\name{mkdate}
\alias{mkdate}
\title{Creates a file path from a name, directory, and extension, then recursively creates each directory in the path.
By default, the path is placed in a directory with today's date.}
\usage{
mkdate(filename, ext, path = ".", append.date = T)
}
\arguments{
\item{filename}{The output file name.}

\item{ext}{The file extension.}

\item{path}{The path to the output file.}

\item{append.date}{Whether to write the output to a directory with today's date.}
}
\value{
The output file path.
}
\description{
Creates a file path from a name, directory, and extension, then recursively creates each directory in the path.
By default, the path is placed in a directory with today's date.
}
\examples{
mkdate("example","txt",path="path/to")
}
